backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.32.3
pandas==2.2.0
numpy==1.26.4
gunicorn==21.2.0
python-dotenv==1.1.0 